ec_enrollee::~ec_enrollee: join pres announcement thread
If Enrollee begins DPP onboarding, it will send presence announcement
frames in a separate thread. If the Enrollee does not receive an
authentication request signaling the thread to join, and the Enrollee
process is killed, the presence announcement frame continues, and the
Configurator continues receiving these frames even though there's no
longer an Enrollee to response to.

This fixes by joining in destructor.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>