RDKBACCL-1161: Fill appropriate Message ID in CMDU header as per spec
This change addresses filling of appropriate Message Id in CMDU header
as per the specification. The responses to the Query or Request
messages to have the same Message ID as that of Query is addressed
as well with this commit.
This commit also addresses correcting the bit field definitions of
structures to be compliant with GCC compiler and also some warning fixes.

Test Procedure:
Tested Basic easymesh scenario of configuration, SSID change and
channel change on RPI.