RDKBACCL-1221: Channel preference report to contain information of all radios.
The channel preference report was being formed for a single radio provided
as part of channel pref query. This commit addresses the change for channel
pref report to contain information of all radios.

Test report: Verified the channel preference query and report
to have appropriate TLVs based on ieee1905 formats